Implement a fix from OpenBSD 5.6-stable:
httpd was developed very rapidly in the weeks before 5.6 release,
and it has a few flaws.  It would be nice to get these flaws fully
remediated before the next release, and that requires the community
to want to use it.
Therefore here is a "jumbo" patch that brings in the most important
fixes.

committing on behalf of reyk@